--- packages/symlink-patcher/.nycrc
@@ -1,1 +1,0 @@
-{"include":["build/src/**"]}
