--- packages/symlink-patcher/prettier.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  singleQuote: true,
-  trailingComma: 'es5',
-};
