--- packages/symlink-patcher/test/test.ts
@@ -1,8 +1,0 @@
-/**
- * @fileoverview Description of this file.
- */
-import * as assert from 'assert';
-
-describe('apples in the sun', () => {
-  assert.ok('dry up');
-});
