--- packages/symlink-patcher/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "./node_modules/gts/tsconfig-google.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "build",
-    "typeRoots":["./node_modules/@types"],
-    "baseUrl":"./"
-  },
-  "include": [
-    "src/**/*.ts",
-    "test/**/*.ts"
-  ]
-}
