--- packages/symlink-patcher/tslint.json
@@ -1,8 +1,0 @@
-{
-  "extends": "gts/tslint.json",
-  "linterOptions": {
-    "exclude": [
-      "**/*.json"
-    ]
-  }
-}
